requirements.txt
flask-login
bcrypt
python-dotenv
psycopg2
flask-wtf